@@ -1,13 +1,13 @@
 #define SGT_VISUALIZATION
 #define SGT_TESTING
 #define CAMERA_DETECTION_CARSTATE
-#define CAMERA_DETECTION_CAMERA_SHOW
+//#define CAMERA_DETECTION_CAMERA_SHOW
 //#define CAMERA_DETECTION_FAKE_LIDAR
 //#define CAMERA_DETECTION_CONSOLE_SHOW
-//#define CAMERA_DETECTION_RECORD_VIDEO
-//#define CAMERA_DETECTION_RECORD_VIDEO_SVO //malo by sa skusit s kamerou
-#define SGT_DEBUG_STATE
-#define SGT_USE_JOYSTICK
+// #define CAMERA_DETECTION_RECORD_VIDEO
+// #define CAMERA_DETECTION_RECORD_VIDEO_SVO
+//#define SGT_DEBUG_STATE
+//#define SGT_USE_JOYSTICK
 //#define FUSION_CONSOLE_SHOW
 //#define SGT_EXPORT_DATA_CSV   // export dat z kalibracie a fuzie
 //#define SIMPLE_FUSION   // popri KF fuzii sa pocita aj jednoducha pre porovnanie
