@@ -369,11 +369,11 @@ void CameraConeDetection::predict(Detector &detector, sl::MODEL &cam_model) {
         m_numOfDetectedCones = result_vec.size();
 #endif
 
-        sgtdv_msgs::ConeArr coneArr;
-        sgtdv_msgs::Cone cone;
-        sgtdv_msgs::Point2D point2D;
+        sgtdv_msgs::ConeStampedArr coneArr;
+        sgtdv_msgs::ConeStamped cone;
+        sgtdv_msgs::Point2DStamped point2D;
 #ifdef CAMERA_DETECTION_FAKE_LIDAR
-        sgtdv_msgs::Point2DArr point2DArr;
+        sgtdv_msgs::Point2DAStampedrr point2DArr;
 #endif//CAMERA_DETECTION_FAKE_LIDAR
         int i_n = 0;
         for (auto &i : result_vec) {
@@ -419,7 +419,7 @@ void CameraConeDetection::predict(Detector &detector, sl::MODEL &cam_model) {
         // std::cout <<""<< std::endl;
 
         // Fill up header message
-        carState.header.stamp = ros::Time::now();
+        carState.header.stamp = capture_time;
         carState.header.frame_id = "odom";
         // Fill up pose message
         carState.pose.pose.position.x = camera_pose.getTranslation().x;
