@@ -9,7 +9,7 @@
 #include "../../SGT_Macros.h"
 #include <sgtdv_msgs/DebugState.h>
 #include <std_msgs/Empty.h>
-#include <sgtdv_msgs/Point2DArr.h>
+#include <sgtdv_msgs/Point2DStampedArr.h>
 #include <sensor_msgs/PointCloud2.h>
 
 int main(int argc, char **argv) {
@@ -17,7 +17,7 @@ int main(int argc, char **argv) {
 
     ros::init(argc, argv, "lidarConeDetection");
     ros::NodeHandle handle;
-    ros::Publisher publisher = handle.advertise<sgtdv_msgs::Point2DArr>("lidar_cones", 1);
+    ros::Publisher publisher = handle.advertise<sgtdv_msgs::Point2DStampedArr>("lidar_cones", 1);
 
 #ifdef SGT_DEBUG_STATE
     ros::Publisher lidarConeDetectionDebugStatePublisher = handle.advertise<sgtdv_msgs::DebugState>("lidar_cone_detection_debug_state", 1);
