@@ -44,27 +44,9 @@ void PathTrackingSynch::StartCallback(const std_msgs::Empty::ConstPtr &msg)
 
 void PathTrackingSynch::Do()
 {
-    while (ros::ok())
-    {        
-        ros::spinOnce();
-
-        if(!m_trajectoryReady || !m_poseReady || !m_velocityReady) continue;
-        
-        auto start = std::chrono::steady_clock::now();
-
+    
+    if(m_trajectoryReady && m_poseReady && m_velocityReady)
+    {
         m_pathTracking.Do(m_pathTrackingMsg);
-
-        auto finish = std::chrono::steady_clock::now();
-        auto timePerFrame = std::chrono::duration_cast<std::chrono::milliseconds>(finish - start).count() / 1000.f;
-        float timeDiff = TIME_PER_FRAME - timePerFrame;
-
-        if (timeDiff > 0.f)
-        {
-            sleep(timeDiff);
-        }
-        else
-        {
-            //defenzivne programovanie ftw
-        } 
     }
 }
